requirements.txt
flask
python-dotenv
pydantic
numpy
scikit-learn
tiktoken
openai>=1.40.0
